Refactor form handling by extracting question rendering and form processing into separate components
